+++ src/spec/rebindEvents.js
@@ -1,0 +1,4 @@
+export function rebindEvent(mediator, event, target, object) {
+  mediator.off(event);
+  mediator.on(event, target, object);
+}