+++ CHANGELOG.md
@@ -7,6 +7,11 @@
 - Move route parameter configuration to appConfig.js.
 - Remove bits and pieces of deprecated code.
 
+## v4.9.0 (2023-10-30)
+
+- Fixed bug with checkboxes not persisting properly when page is reloaded with
+  facet selections in the URL
+
 ## v4.8.1 (2023-10-20)
 
 - Minor bug fix with the release script